requirements.txt
flask==2.0.1
twilio==7.0.0
requests==2.26.0
python-dotenv==0.19.0
pytest==7.4.0
pytest-cov==4.1.0 